+++ taxsystem/models/filters.py
@@ -2,12 +2,14 @@
 
 # Django
 
+# Django
 from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
 from django.db import models
 from django.db.models import Q
 from django.utils.translation import gettext_lazy as _
 
+# AA TaxSystem
 from taxsystem.models.tax import OwnerAudit, Payments
 
 
@@ -110,8 +112,8 @@ class SmartGroup(models.Model):
     class Meta:
         default_permissions = ()
 
-    corporation = models.OneToOneField(
-        OwnerAudit, on_delete=models.CASCADE, related_name="filter_sets"
+    owner = models.OneToOneField(
+        OwnerAudit, on_delete=models.CASCADE, related_name="ts_filter_sets"
     )
     description = models.CharField(max_length=255)
     name = models.CharField(max_length=100)