+++ taxsystem/signals.py
@@ -1,12 +1,17 @@
-import logging
-
+# Django
 from django.db.models.signals import post_save, pre_delete
 
+# Alliance Auth
 from allianceauth import hooks
+from allianceauth.services.hooks import get_extension_logger
 
-from taxsystem import models
+# Alliance Auth (External Libs)
+from app_utils.logging import LoggerAddTag
 
-logger = logging.getLogger(__name__)
+# AA TaxSystem
+from taxsystem import __title__, models
+
+logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 class HookCache: